Python Fetch/requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
beautifulsoup4 >= 4.14.3
mysql-connector-python >= 9.5.0
requests >= 2.32.5